+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules/
 npm-debug.log
-controller/js/build/
+dashboard/js/build/
