+++ dashboard/index.html
@@ -4,8 +4,7 @@
   <meta charset="UTF-8">
   <title></title>
   <script type="text/javascript" src="/socket.io/socket.io.js"></script>
-  <script type="text/javascript" src="main.js"></script>
-  <link rel="stylesheet" href="style.css" />
+  <script type="text/javascript" src="js/build/bundle.js"></script>
 </head>
 <body>
   <header>
