+++ package.json
@@ -4,10 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "build": "gulp webpack",
-    "build:min": "gulp webpack --min",
-    "build:watch": "gulp webpack --watch",
-    "build:browser-sync": "gulp webpack --browser-sync"
+    "build": "webpack --progress --colors",
+    "build:watch": "webpack --progress --colors --watch"
   },
   "repository": {
     "type": "git",
@@ -19,7 +17,16 @@
     "url": "https://github.com/shundroid/sphero-joystick/issues"
   },
   "homepage": "https://github.com/shundroid/sphero-joystick#readme",
-  "devDependencies": {},
+  "devDependencies": {
+    "babel-core": "^6.10.4",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.9.0",
+    "css-loader": "^0.23.1",
+    "gulp": "^3.9.1",
+    "minimist": "^1.2.0",
+    "style-loader": "^0.13.1",
+    "webpack": "^1.13.1"
+  },
   "dependencies": {
     "argv": "0.0.2",
     "express": "^4.14.0",
